refactor: 테스트 클래스들을 적절한 패키지로 이동 (domain, service)
