+++ src/test/java/com/kakao/onboarding/precourse/albusduke/lotto/domain/PrizeTest.java
@@ -1,4 +1,4 @@
-package com.kakao.onboarding.precourse.albusduke.lotto;
+package com.kakao.onboarding.precourse.albusduke.lotto.domain;
 
 import com.kakao.onboarding.precourse.albusduke.lotto.domain.GameResult;
 import com.kakao.onboarding.precourse.albusduke.lotto.domain.Prize;
